+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+ - Add calculated totals for QSI as the totals aren't passed to SDX
 
 ### 3.10.0 2019-04-02
  - Update Jinja to 2.10.1 to fix a security issue.